requirements.txt
flask
flask_session
pysyslogclient
mysql-connector-python
pymysql
